client/requirements.txt
matplotlib==3.7.1
Pillow==9.5.0
requests==2.28.2
